Separate console output functions into their own module & track amendments.
