+++ process_responses/process_validation_output.py
@@ -6,74 +6,6 @@ import time
 import asyncio
 from copy import deepcopy
 
-from prettytable import PrettyTable
-
-from .xrpl_version_decoder import decode_version
-
-async def format_table_validation(table):
-    '''
-    Format output for the validation table, so it's human friendly.
-
-    :param list table: Dictionaries for each validator being tracked
-    '''
-    color_reset = "\033[0;0m"
-    green = "\033[0;32m"
-    red = "\033[1;31m"
-    for validator in table:
-        if isinstance(validator['master_key'], str):
-            validator['master_key'] = validator['master_key'][:5]
-        if isinstance(validator['validation_public_key'], str):
-            validator['validation_public_key'] = validator['validation_public_key'][:5]
-        if isinstance(validator['ledger_hash'], str):
-            validator['ledger_hash'] = validator['ledger_hash'][:5]
-        if isinstance(validator['validated_hash'], str):
-            validator['validated_hash'] = validator['validated_hash'][:5]
-        if validator['forked'] is False:
-            validator['forked'] = green + str(validator['forked']) + color_reset
-        else:
-            validator['forked'] = red + str(validator['forked']) + color_reset
-        if validator['full']:
-            validator['full'] = green + str(validator['full']) + color_reset
-        else:
-            validator['full'] = red + str(validator['full']) + color_reset
-        if isinstance(validator['server_version'], str):
-            if validator['server_version'][0:].isdigit():
-                server_version = await decode_version(validator['server_version'])
-                validator['server_version'] = server_version.get('version')
-    return table
-
-async def print_table_validation(table):
-    '''
-    Print the validation table.
-
-    :param list table: Dictionaries for each validator being tracked
-    '''
-    logging.info("Preparing to print updated validations table.")
-    pretty_table = PrettyTable()
-    pretty_table.field_names = [
-        "Validator Name", "Master Key", "Eph Key", "Version", "Base Fee", "Local LL Fee",
-        "LL Hash", "LL Index", "Full?", "Forked?", "Last Updated",
-    ]
-
-    table_new = await format_table_validation(deepcopy(table))
-
-    for validator in table_new:
-        pretty_table.add_row([
-            validator['server_name'],
-            validator['master_key'],
-            validator['validation_public_key'],
-            validator['server_version'],
-            validator['base_fee'],
-            validator['load_fee'],
-            validator['ledger_hash'],
-            validator['ledger_index'],
-            validator['full'],
-            validator['forked'],
-            validator['time_updated'],
-        ])
-
-    print(pretty_table)
-    logging.info("Successfully printed updated validations table.")
 
 async def del_dup_validators(table):
     '''
